test: file reorg (rename + merges)
Pure organizational reshuffle. No tests added or removed; pre/post count
434 passed. Each move addresses a naming inversion or arbitrary file split
that was making the test directory harder to navigate.

Rename:
- tests/test_trajectory.py → tests/test_atif_trajectory.py
  test_trajectory.py tested the unwired trajectories/ ATIF package, not the
  production _trajectory.py capture path (which is tested by
  test_capture_trajectory.py). The old name was an inversion. Renamed to
  match the package it actually exercises.

Merges:
- tests/test_credentials.py + tests/test_env_mapping.py → tests/test_agent_registry.py
  Both test AgentConfig/ProviderConfig registry shape. git mv on
  test_credentials.py preserves history; test_env_mapping.py content
  appended verbatim, no test bodies modified.

- tests/test_dep_staging.py → tests/test_env_setup.py
  Both already imported from benchflow._env_setup. The split was
  historically arbitrary. Appended TestDepLocalName + TestStageDockerfileDeps
  verbatim and added _dep_local_name + stage_dockerfile_deps to the existing
  import block.

- tests/test_exclude_tasks.py::TestNativeYamlNewFields → tests/test_yaml_config.py
  TestNativeYamlNewFields tests Job.from_yaml parsing of exclude/agent_env/
  sandbox_user fields — that belongs alongside the other YAML config tests,
  not in the Job filter file. Helper _make_tasks copied (small, self-contained,
  not worth a conftest extraction). test_exclude_tasks.py is now narrowly
  scoped to TestExcludeTasksFilter; updated its module docstring to match.
