+++ tests/test_agent_registry.py
@@ -1,5 +1,7 @@
-"""Tests for AgentConfig.env_mapping — translates BENCHFLOW_PROVIDER_* to agent-native vars."""
+"""Tests for AgentConfig + ProviderConfig registry shape:
+env_mapping (BENCHFLOW_PROVIDER_* → agent-native vars) and credential_files."""
 
+from benchflow.agents.providers import PROVIDERS
 from benchflow.agents.registry import AGENTS, AgentConfig
 
 
@@ -36,3 +38,39 @@ class TestEnvMappingField:
         """Openclaw reads BENCHFLOW_PROVIDER_* directly via shim — no mapping needed."""
         cfg = AGENTS["openclaw"]
         assert cfg.env_mapping == {}
+
+
+class TestAgentCredentialFiles:
+    def test_codex_has_auth_json(self):
+        cfg = AGENTS["codex-acp"]
+        assert len(cfg.credential_files) == 1
+        cf = cfg.credential_files[0]
+        assert cf.env_source == "OPENAI_API_KEY"
+        assert ".codex/auth.json" in cf.path
+        assert "{home}" in cf.path
+        assert "{value}" in cf.template
+
+    def test_claude_no_credential_files(self):
+        cfg = AGENTS["claude-agent-acp"]
+        assert cfg.credential_files == []
+
+    def test_openclaw_no_credential_files(self):
+        cfg = AGENTS["openclaw"]
+        assert cfg.credential_files == []
+
+
+class TestProviderCredentialFiles:
+    def test_vertex_providers_have_adc(self):
+        for name in ("google-vertex", "anthropic-vertex"):
+            cfg = PROVIDERS[name]
+            assert len(cfg.credential_files) == 1, (
+                f"{name} should have 1 credential_file"
+            )
+            cf = cfg.credential_files[0]
+            assert cf["env_source"] == "GOOGLE_APPLICATION_CREDENTIALS_JSON"
+            assert "gcloud" in cf["path"]
+            assert "GOOGLE_APPLICATION_CREDENTIALS" in cf.get("post_env", {})
+
+    def test_zai_no_credential_files(self):
+        cfg = PROVIDERS["zai"]
+        assert cfg.credential_files == []
