+++ tests/test_exclude_tasks.py
@@ -1,4 +1,4 @@
-"""Tests for exclude_tasks, agent_env, and sandbox_user in Job config."""
+"""Tests for the exclude_tasks filter in Job._get_task_dirs."""
 
 from benchflow.job import Job, JobConfig
 
@@ -41,53 +41,3 @@ class TestExcludeTasksFilter:
         job = Job(tasks_dir=tasks_dir, jobs_dir=tmp_path / "jobs", config=cfg)
         dirs = job._get_task_dirs()
         assert len(dirs) == 3
-
-
-class TestNativeYamlNewFields:
-    def test_exclude_parsed(self, tmp_path):
-        _make_tasks(tmp_path)
-        config = tmp_path / "config.yaml"
-        config.write_text("""
-tasks_dir: tasks
-exclude:
-  - task-a
-  - task-c
-""")
-        job = Job.from_yaml(config)
-        assert job._config.exclude_tasks == {"task-a", "task-c"}
-        dirs = job._get_task_dirs()
-        assert [d.name for d in dirs] == ["task-b"]
-
-    def test_agent_env_parsed(self, tmp_path):
-        _make_tasks(tmp_path)
-        config = tmp_path / "config.yaml"
-        config.write_text("""
-tasks_dir: tasks
-agent_env:
-  MY_KEY: my-value
-  OTHER_KEY: other-value
-""")
-        job = Job.from_yaml(config)
-        assert job._config.agent_env == {
-            "MY_KEY": "my-value",
-            "OTHER_KEY": "other-value",
-        }
-
-    def test_sandbox_user_parsed(self, tmp_path):
-        _make_tasks(tmp_path)
-        config = tmp_path / "config.yaml"
-        config.write_text("""
-tasks_dir: tasks
-sandbox_user: testuser
-""")
-        job = Job.from_yaml(config)
-        assert job._config.sandbox_user == "testuser"
-
-    def test_defaults_when_omitted(self, tmp_path):
-        _make_tasks(tmp_path)
-        config = tmp_path / "config.yaml"
-        config.write_text("tasks_dir: tasks\n")
-        job = Job.from_yaml(config)
-        assert job._config.exclude_tasks == set()
-        assert job._config.agent_env == {}
-        assert job._config.sandbox_user == "agent"
